requirements.txt
PySide6
spacy
pymupdf
python-docx
